Refactored computations
